+++ google-cloud-tools-plugin/src/com/google/cloud/tools/intellij/appengine/cloud/standard/AppEngineStandardStagingPropertiesPanel.java
@@ -143,7 +143,7 @@ public class AppEngineStandardStagingPropertiesPanel {
   }
 
   @VisibleForTesting
-  public JTextField getJarSplittingExcludesTextField() {
+  public JBTextField getJarSplittingExcludesTextField() {
     return jarSplittingExcludesTextField;
   }
